requirements.txt
Django==5.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
PyMySQL==1.1.0
minio==7.2.0
elasticsearch==8.11.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
hanziconv==0.3.2
nltk==3.8.1
datrie==0.8.2 